+++ previous-versions/files/RELEASES.md
@@ -637,3 +637,13 @@ Based on:
 - [typescript v0.38.1] previous-versions/files
 ### Releases
 - [NPM v0.38.1] https://www.npmjs.com/package/@codat/files/v/0.38.1 - previous-versions/files
+
+## 2023-10-03 00:21:19
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Files.yaml
+- Speakeasy CLI 1.93.1 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.38.2] previous-versions/files
+### Releases
+- [NPM v0.38.2] https://www.npmjs.com/package/@codat/files/v/0.38.2 - previous-versions/files